Add new results using scaled irradiance
Irradiance was scaled by setting 700 umol/m2/s equal to 100 mmol/gDW/h,
so 200 umol/m2/s corresponds to 286 mmol/gDW/h.
The lower bound for biomass flux was decreased to 40% of the optimum
to arrive at feasible problems.
The ressults are re-structured into sub-folders that correspond to the
method used to model photon uptake.